Объединил сетевые сервисы в одном классе
