+++ FakeNFT/Catalog/Models/Collection.swift
@@ -5,12 +5,11 @@
 //  Created by Сергей Денисенко on 27.03.2024.
 //
 
-import Foundation
-
+import UIKit
 struct Collection {
     let createdAt: Date?
     let name: String
-    let cover: String
+    let cover: UIImage?
     let nfts: [String]
     let description: String
     let author: String
