+++ FakeNFT/Catalog/NetWork/NetworkError.swift
