+++ FakeNFT/Catalog/Protocols/CatalogPresenterProtocol.swift
@@ -4,11 +4,11 @@
 //
 //  Created by Сергей Денисенко on 27.03.2024.
 //
-import UIKit
+import Foundation
 protocol CatalogPresenterProtocol {
     var view: CatalogViewControllerProtocol? { get set }
     func getValueCount() -> Int
-    func getImagesForCell(index: Int) -> UIImageView?
+//    func getImagesForCell(index: Int) -> Collection?
     func getLabelText(index: Int) -> String
     func viewDidLoad()
     func showSortAlert()
