+++ FakeNFT/Catalog/Protocols/CollectionScreenPresenterProtocol.swift
@@ -6,12 +6,11 @@
 //
 
 import Foundation
-import Kingfisher
 protocol CollectionScreenPresenterProtocol {
 //    var view: CatalogViewControlledProtocol? { get set }
     func getValueCount() -> Int
-    func getOptions() -> (urlCover: URL, options: KingfisherOptionsInfo)?
-    func getMainInfo() -> (name: String, author: String, description: String)
+    func getOptions() -> Collection
+//    func getMainInfo() -> (name: String, author: String, description: String)
     func viewDidLoad()
 //    func showSortAlert()
 }
